pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pytorch-lightning-template"
dynamic = ["version"]
description = "A sample Pytorch Lightning project"
readme = "README.md"

requires-python = ">=3.11"
license = { file = "LICENSE.txt" }
keywords = ["sample", "setuptools", "development"]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    'numpy<2.0.0',
    'opencv-python<5.0.0',
    'torch~=2.3.0',
    'transformers~=4.40.2',
    'albumentations~=1.4.13',
    'pytorch-lightning~=2.3.0',
    'torchvision~=0.18.0',
    'wandb~=0.17.1',
    'ruamel.yaml~=0.18.6',
    'pycocotools~=2.0.8'
]
[project.optional-dependencies]
test = ['pytest']
analyze = ['pyright', 'pylint', 'bandit', 'black', 'isort']
build = ['setuptools', 'wheel', 'build']
publish = ['twine']
all = ["pytorch-lightning-template[test,analyze,build,publish]"]

[tool.setuptools.dynamic]
version = { file = "VERSION" }


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pytorch_lightning_template = ["py.typed"]